[BUGFIX] ViewHelperDocumentation fails with DI
Class ViewHelperDocumentation instantiates the
ViewHelper to retrieve argument information.
This fails when VH's need __construct() arguments,
which are typically DI related.
We however only need the arguments array, and
this has typically nothing to do with DI.
The simple solution is to reflect the class
and instantiate it without calling the constructor.

Resolves #12