[TASK] Add missing ViewHelper argument data type notation "bool" (#16)
According to `TYPO3Fluid\Fluid\Core\Parser\TemplateParser`, in addition to the data type `boolean`, `bool` is also treated as a boolean value. To prevent the ViewHelper documentation from outputting the data type `mixed` if the string `bool` was specified as a data type in the ViewHelper, it is necessary to adapt the XSD generation accordingly.